refractor(ART12300): move avatar initials logic to src/utils
